Cambio de tipo de variables
Antes, en el endpoint del filtro múltiple, las variables diametro, ancho y alto eran Integers, pero debian permitir decimales, así que pasó a double. Además de actualizar el README.md

+++ src/main/java/com/TFCFelipePerezTevar/springbootTFCFelipePerezTevar/controller/ProductoColorController.java
@@ -33,8 +33,8 @@ public class ProductoColorController {
 
 
     @GetMapping("/filtro")
-    public ResponseEntity<List<Productocolor>> listFiltro2(@RequestParam("diametro") Integer diametro, @RequestParam("alto")Integer alto,
-                                                           @RequestParam("ancho") Integer ancho, @RequestParam("tipo") String tipo, @RequestParam("modelo") String modelo){
+    public ResponseEntity<List<Productocolor>> listFiltro2(@RequestParam("diametro") Double diametro, @RequestParam("alto")Double alto,
+                                                           @RequestParam("ancho") Double ancho, @RequestParam("tipo") String tipo, @RequestParam("modelo") String modelo){
 
         if(diametro == 0){
             diametro = null;
